+++ serviceTest/src/test/scala/org/broadinstitute/dsde/workbench/service/Orchestration.scala
@@ -322,18 +322,39 @@ trait Orchestration extends RestClient with LazyLogging with SprayJsonSupport wi
       postRequest(apiUrl(s"api/methods/$ns/$name/$snapshotId/permissions"), request)
     }
   }
+
   /*
-     *  NIH requests
-    */
-    object NIH {
-      def addUserInNIH(jwt: String)(implicit token: AuthToken): Unit = {
+   *  NIH requests
+   */
+  object NIH {
+
+    case class NihStatus(linkedNihUsername: Option[String] = None,
+                         datasetPermissions: Set[NihDatasetPermission],
+                         linkExpireTime: Option[Long] = None)
+
+    case class NihDatasetPermission(name: String, authorized: Boolean)
+
+    implicit val impNihDatasetPermission = jsonFormat2(NihDatasetPermission)
+    implicit val impNihStatus = jsonFormat3(NihStatus.apply)
+
+    def addUserInNIH(jwt: String)(implicit token: AuthToken): Unit = {
       logger.info(s"Adding user to NIH whitelist: $jwt")
       postRequest(apiUrl(s"/api/nih/callback"), Map("jwt" -> jwt))
     }
 
-      def refreshUserInNIH(jwt: String)(implicit token: AuthToken): Unit = {
-        logger.info(s"Refershing user's NIH status")
-        NIH.addUserInNIH(jwt)
+    def refreshUserInNIH(jwt: String)(implicit token: AuthToken): Unit = {
+      logger.info(s"Refreshing user's NIH status")
+      NIH.addUserInNIH(jwt)
+    }
+
+    def syncWhitelistFull()(implicit token: AuthToken): Unit = {
+      logger.info(s"Syncing whitelist (full)")
+      postRequest(apiUrl("/sync_whitelist"))
+    }
+
+    def getUserNihStatus()(implicit token: AuthToken): NihStatus = {
+      val response = getRequest(apiUrl("/api/nih/status"))
+      parseResponse(response).parseJson.convertTo[NihStatus]
     }
 
   }